requirements.txt
Django>=5.0.3
pdfplumber
openai>=1.68.2,<2.0.0
python-dotenv
langchain-core
langchain-openai
langchain-text-splitters
langchain-community
faiss-cpu
pandas
selenium
webdriver-manager
matplotlib
seaborn
beautifulsoup4